feat: 迁移算力源注册&国际化
