+++ ui/src/locales/en-US/login.ts
@@ -2,7 +2,7 @@
  * 登录页面 国际化配置
  */
 export default {
-  'pages.login.title': 'NascentCore.AI Cloud',
+  'pages.login.title': 'NascentCore Cloud',
   'pages.login.username': 'Please enter your email',
   'pages.login.password': 'Please enter your password',
   'pages.login.rememberMe': 'Remember me',
@@ -12,12 +12,16 @@ export default {
   'pages.login.serviceEmail': 'Service email',
 
   'pages.regist.title': 'User Registration',
+  'pages.regist.type.user': 'User registration',
+  'pages.regist.type.source': 'Source registration',
   'pages.regist.username': 'Please enter your email',
   'pages.regist.codemes': 'Please enter the email verification code',
   'pages.regist.codemes.tip': 'Resend in seconds',
-  'pages.regist.codemes.tip2': 'Get code',
+  'pages.regist.codemes.tip2': 'Get verification code',
   'pages.regist.password': 'Please enter a password',
   'pages.regist.password2': 'Confirm password',
+  'pages.regist.companyName': 'Please enter the company name',
+  'pages.regist.companyPhone': 'Please enter the contact phone number',
   'pages.regist.password.error': 'The passwords entered do not match',
   'pages.regist.agreeTerms.error': 'Please agree to the Suanxiang Cloud Privacy Policy',
   'pages.regist.agreeTerms': 'Please agree to the Suanxiang Cloud Privacy Policy',
