+++ ui/src/locales/zh-CN/login.ts
@@ -12,12 +12,16 @@ export default {
   'pages.login.serviceEmail': '客服邮箱',
 
   'pages.regist.title': '用户注册',
+  'pages.regist.type.user': '算力用户注册',
+  'pages.regist.type.source': '算力源注册',
   'pages.regist.username': '请输入邮箱',
   'pages.regist.codemes': '请输入邮箱验证码',
   'pages.regist.codemes.tip': '秒后重新发送',
   'pages.regist.codemes.tip2': '获取验证码',
   'pages.regist.password': '请输入密码',
   'pages.regist.password2': '请确认密码',
+  'pages.regist.companyName': '请输入公司名称',
+  'pages.regist.companyPhone': '请输入联系电话',
   'pages.regist.password.error': '两次输入的密码不一致',
   'pages.regist.agreeTerms.error': '请同意算想未来隐私政策',
   'pages.regist.agreeTerms': '请同意算想未来隐私政策',
